+++ src/Filters/jsonstor-userinfo.js
@@ -3,7 +3,7 @@
 
 const LIB_UUID = require( 'uuid' );
 
-const jsongin = require( '@liquicode/jsongin' )();
+const jsongin = require( '@liquicode/jsongin' );
 // const jsonstor = require( '../jsonstor' )();